# docs/swagger.yaml
definitions:
...
  models.CommentVoteRequest:
    description: Request model for voting on a comment
    properties:
      vote_type:
        allOf:
        - $ref: '#/definitions/models.VoteType'
        example: 1
    required:
    - vote_type
    type: object
  models.CommentVoteResponse:
    description: Response model for a comment vote operation
    properties:
      comment_id:
        example: 1
        type: integer
      upvote_count:
        example: 5
        type: integer
      user_vote:
        example: 1
        type: integer
    type: object
...
  models.DeleteFileRequest:
    description: Request model for deleting a file
    properties:
      file_url:
        example: https://example.com/file.jpg
        type: string
    type: object
  models.FileUploadResponse:
    description: Response model for editor file upload
    properties:
      file_url:
        example: https://example.com/file.jpg
        type: string
    type: object
...
  models.StandardResponse:
    description: A standard API response format with data payload
    properties:
      data: {}
      error:
        example: Invalid input
        type: string
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
...
  models.VoteType:
    enum:
    - -1
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - VoteTypeDownvote
    - VoteTypeNone
    - VoteTypeUpvote
info:
  contact: {}
paths:
...
  /comments/{commentID}/vote:
    post:
      consumes:
      - application/json
      description: Upvote, downvote, or remove a vote from a comment
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      - description: Vote type
        in: body
        name: vote
        required: true
        schema:
          $ref: '#/definitions/models.CommentVoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated vote counts
          schema:
            $ref: '#/definitions/models.CommentVoteResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Vote on a comment
      tags:
      - Comments
  /comments/{commentID}/votes:
    get:
      description: Returns the upvote/downvote counts for a specific comment
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Vote counts
          schema:
            $ref: '#/definitions/models.CommentVoteResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Get vote counts for a comment
      tags:
      - Comments
  /files/delete:
    post:
      consumes:
      - application/json
      description: Delete a file that was previously uploaded for use in the editor
      parameters:
      - description: File URL to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: File deleted successfully
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete a file uploaded for editor use
      tags:
      - Files
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file that can be used in the editor when creating or editing
        posts
      parameters:
      - description: File to upload (JPG, JPEG, PNG, WEBP, GIF, SVG, PDF, max 5MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded successfully
          schema:
            $ref: '#/definitions/models.FileUploadResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Upload a file for editor use
      tags:
      - Files
...
swagger: "2.0"